+++ src/hocks/useMap.ts
@@ -2,7 +2,7 @@
 import {useEffect, useState, MutableRefObject, useRef} from 'react';
 import {Map, TileLayer} from 'leaflet';
 import 'leaflet/dist/leaflet.css';
-import { CityType } from '../mocks/offers';
+import { CityType } from '../types/types';
 
 
 function useMap(
