+++ src/mocks/offers.ts
@@ -1,36 +1,13 @@
-
-export type CityType = {
-  name: string;
-  location: {
-    lat: number;
-    lng: number;
-    zoom: number;
-  };
-};
-
-export type OfferType = {
-  id: string;
-  mark: string;
-  src: string;
-  price: number;
-  description: string;
-  type: string;
-  city: CityType;
-  location: {
-    lat: number;
-    lng: number;
-    zoom: number;
-  };
-};
+import { OfferType } from '../types/types';
 
 export const offersMock: OfferType[] = [
   {
     id: 'b0e9d6ef-8b5b-465c-8cf6-bb52a268f4ba',
-    mark: 'Basic',
     src: 'img/apartment-01.jpg',
     price: 70,
     description: 'Beautiful & luxurious apartment at great location',
     type: 'Apartment',
+    isPremium: false,
     city: {
       name: 'Amsterdam',
       location: {
@@ -47,11 +24,11 @@ export const offersMock: OfferType[] = [
   },
   {
     id: '6d15ce78-fad6-45a0-812c-045e5ca5cb76',
-    mark: 'Premium',
     src: 'img/apartment-02.jpg',
     price: 170,
     description: 'Beautiful & luxurious hotel at great location',
     type: 'Hotel',
+    isPremium: false,
     city: {
       name: 'Amsterdam',
       location: {
@@ -68,11 +45,11 @@ export const offersMock: OfferType[] = [
   },
   {
     id: '0f4cdf1f-c954-4742-874b-b666ec6d8015',
-    mark: 'Business',
     src: 'img/apartment-03.jpg',
     price: 100,
     description: 'Beautiful & luxurious apartment at great location',
     type: 'Apartment',
+    isPremium: false,
     city: {
       name: 'Amsterdam',
       location: {
@@ -89,11 +66,11 @@ export const offersMock: OfferType[] = [
   },
   {
     id: '8e5fe338-06d7-4e80-b7a1-09d52c100cf6',
-    mark: 'Comfort',
     src: 'img/apartment-01.jpg',
     price: 90,
     description: 'Beautiful & luxurious hotel at great location',
     type: 'Hotel',
+    isPremium: true,
     city: {
       name: 'Amsterdam',
       location: {
@@ -113,11 +90,11 @@ export const offersMock: OfferType[] = [
 export const offersNearbyMock: OfferType[] = [
   {
     id: 'b0e9d6ef-8b5b-465c-8cf6-bb52a268f4ba',
-    mark: 'Basic',
     src: 'img/apartment-01.jpg',
     price: 70,
     description: 'Beautiful & luxurious apartment at great location',
     type: 'Apartment',
+    isPremium: true,
     city: {
       name: 'Amsterdam',
       location: {
@@ -134,11 +111,11 @@ export const offersNearbyMock: OfferType[] = [
   },
   {
     id: '6d15ce78-fad6-45a0-812c-045e5ca5cb76',
-    mark: 'Premium',
     src: 'img/apartment-02.jpg',
     price: 170,
     description: 'Beautiful & luxurious hotel at great location',
     type: 'Hotel',
+    isPremium: false,
     city: {
       name: 'Amsterdam',
       location: {
@@ -155,11 +132,11 @@ export const offersNearbyMock: OfferType[] = [
   },
   {
     id: '0f4cdf1f-c954-4742-874b-b666ec6d8015',
-    mark: 'Business',
     src: 'img/apartment-03.jpg',
     price: 100,
     description: 'Beautiful & luxurious apartment at great location',
     type: 'Apartment',
+    isPremium: false,
     city: {
       name: 'Amsterdam',
       location: {
